Reorganize code
- Move item.historical.OpenKAPSARC etc. to item.openkapsarc.OpenKAPSARC,
  since it will eventually be used by the model database also.
- Promote @TheYuanLiao's item.historical.utils to item.historical.
- Move mapping_conv_phase1.yaml in with other metadata files.
- Adjust CLI etc.
